Split test/lib/server-reservation.test.ts, drop its biome override
Removed test/lib/server-reservation.test.ts (1224 lines) from biome.json's
noExcessiveLinesPerFile override list and split it into five topic files
under test/lib/server-reservation/, each well under 400 lines:
deposit-basics, public-default-modifiers, promo-addons, edge-cases, and
no-provider, plus a shared helpers.ts.

Splitting surfaced 19 jscpd clones; fixed by extracting shared helpers into
helpers.ts rather than ignoring the duplication:
- stubCheckout / captureCheckoutIntent for the repeated
  stub-checkout-provider-and-capture-the-intent flow
- submitBuyerOrder for the common quantity/email/name form submission
- setupReservationListing for the setupStripe + booking fee + reservation
  amount + listing preamble shared across checkout/webhook tests
- createOptionalAddOn, addServiceCharge, createProgrammeCharge,
  createSave10Promo for repeated modifier fixtures
- expectRefundedPlaceholder for the shared "kept as a refunded quantity-0
  placeholder" assertion tail
